Create Delete-Recreate-WifiService
Basic script for deleting/recreating the Wi-Fi Network service in MacOS 13 (Ventura) to MacOS 14.3.1 (Sonoma)

Context: I encountered an issue where a MacBook would successfully connect to an iPhone's (iOS 17) mobile hotspot, but not be able to successfully connect out to the internet. 

The easiest/most immediate fix was to delete the Wi-Fi service on the Macbook and recreate it. After doing this and reconnecting to an iPhone's mobile hotspot, the MacBook would have full internet connectivity.